requirements.txt
PyMuPDF
Pillow
openai
fastapi
uvicorn
python-multipart
numpy